Change behavior of FixedPriceBenefit to be more logical
Oscar's default FixedPriceBenefit is unintuitive. It ignores the benefit range and
the max_affected_items and uses the products affected by the condition instead. This
changes the behavior to more closely follow how other benefits work. It applied, it
gives the basket items in the benefit range for a fixed price, not the basket items
in the condition range. It also respects the max_affected_items setting.