Fix bug in condition line consumption
Fix a bug that occurred when the condition and the benefit of an offer had
different ranges. The condition wouldn't consume the triggering products in some
cases, which lead to a single product being able to trigger discounting of
multiple other lines.

This change fixes the bug and ensures that any line triggering a discount (even
on another line) is consumed so that it can not trigger any other discounts
within an offer group.