Refs #10566 Alter behavior of MultibuyDiscountBenefit
Previously the MultibuyDiscountBenefit had the functionality to give the
cheapest product for free. This commit changes the behavior to be more consumer
friendly and to better match consuimer expectations of thow this benefit should
work.

The new behavior is to give the second most expensive product in the selected
range for free. The second-most-expensive because the idea is that the
first-most-expensive will be consumed by a CountCondition.

In other words, the following rules apply:

1. If two products in cart, highest product consumer pays for and the
    lowest price item is given for free
2. If three products in cart, highest product consumer pays for, second
    highest should be free, third the consumer should pay for
3. If four products are in the cart, consumer should pay full price for
    the highest price, then the second highest item should be free.
    Then the third highest consumer pays for, then the fourth item
    would be free.